feat(sidebar,home): move Continue practicing to a desktop rail card, drop the home section
The 16:9 hero dominated the desktop home page, and on mobile the
last-practiced item is already the first row of the recents grid — so
Home carries no Continue section on any platform now. The desktop rail
gets a compact card above the account chip instead:

- SidebarContinuePracticeCard: thumbnail strip + label/title/meta/
  progress, opens the player via the same warm-and-resume path as the
  old hero; shrinks away when there is nothing to resume
- delete ContinuePracticeCard and its test; the label helpers move into
  the sidebar card file
- rename home_continue_practice_provider -> continue_practice_provider
  (homeContinuePracticeProvider -> continuePracticeResumeProvider); it
  no longer has anything to do with Home
- root_shell_test pins the resume provider off: the card's drift watch
  streams tripped flutter_test's pending-timer guard at teardown, and
  the shell layout tests never exercise resume logic (card is covered
  in app_sidebar_test, including position above the account chip)
